+++ internal/conn/grpc_client_stream.go
@@ -99,10 +99,6 @@ func (s *grpcClientStream) SendMsg(m interface{}) (err error) {
 			return xerrors.WithStackTrace(err)
 		}
 
-		defer func() {
-			s.parentConn.onTransportError(ctx, err)
-		}()
-
 		if !s.wrapping {
 			return err
 		}
@@ -131,7 +127,7 @@ func (s *grpcClientStream) finish(err error) {
 	s.streamCancel()
 }
 
-func (s *grpcClientStream) RecvMsg(m interface{}) (err error) { //nolint:funlen
+func (s *grpcClientStream) RecvMsg(m interface{}) (err error) {
 	var (
 		ctx    = s.streamCtx
 		onDone = trace.DriverOnConnStreamRecvMsg(s.parentConn.config.Trace(), &ctx,
@@ -157,10 +153,6 @@ func (s *grpcClientStream) RecvMsg(m interface{}) (err error) { //nolint:funlen
 		if xerrors.IsContextError(err) {
 			return xerrors.WithStackTrace(err)
 		}
-
-		defer func() {
-			s.parentConn.onTransportError(ctx, err)
-		}()
 
 		if !s.wrapping {
 			return err