+++ internal/conn/state/state_test.go
@@ -1,4 +1,4 @@
-package conn
+package state
 
 import (
 	"testing"