+++ internal/table/session.go
@@ -19,6 +19,7 @@ import (
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/metadata"
 
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/balancer"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/conn"
 	balancerContext "github.com/ydb-platform/ydb-go-sdk/v3/internal/endpoint"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/feature"
@@ -277,7 +278,8 @@ func newSession(ctx context.Context, cc grpc.ClientConnInterface, config *config
 func newTableSession(
 	ctx context.Context, cc grpc.ClientConnInterface, config *config.Config,
 ) (*Session, error) {
-	response, err := Ydb_Table_V1.NewTableServiceClient(cc).CreateSession(ctx,
+	response, err := Ydb_Table_V1.NewTableServiceClient(cc).CreateSession(
+		balancer.BanOnOperationError(ctx, Ydb.StatusIds_OVERLOADED),
 		&Ydb_Table.CreateSessionRequest{
 			OperationParams: operation.Params(
 				ctx,