+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Added pessimization of connection to YDB node when `OVERLOADED` operation error is returned on `CreateSession` request for exclude next `CreateSession` calls on overloaded nodes
+
 ## v3.127.6
 * Added `pkg/xslices.Subtract` helper for subtract elements from source slice with order preservation